Clean packages, move classes
