+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/classic/HypixelLegacy.java
@@ -1,4 +1,4 @@
-package io.github.hypixel_api_wrapper.wrapper.games.legacy;
+package io.github.hypixel_api_wrapper.wrapper.games.classic;
 
 import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
